Handle linked docker container ENV var issues
When linking a container, such as consul or etc to the RabbitMQ container,  Docker will create an OS environment variable that conflicts with configured values. This change makes it so that if the docker linked environment variable is set for a port (CONSUL_PORT, ETCD_PORT), the port will be parsed out of it.

- Add a mapped column "is_port"
- Add a new port parser that will parse out the port of what is specified in port env vars if they're set as URIs