Add more test coverage for autocluster_consul
A bug crept into 0.5.0 with regard to the building/appending of ACL tokens (#60). These tests get the module to 90% coverage, including the almost all the parts that have changed.

We should still push to get to 100% coverage of the module, as the remaining areas largely have to do with parsing of the payload that is returned when nodelist() is invoked and with building the payload that is sent during registration.

Coverage would have been helpful here since the things that broker were introduced in pull requests that were merged that had there been test coverage, the tests would not have passed.