Fix ACL usage (#60)
A bug existed in the appending of access tokens to the query args, where it was being appended as a nested proplist, instead of as a tuple in the query arg list. This became a problem when trying to build the query arg string in autocluster_httpc when the datastructure of the query arg list was not correct. Tests have been added to autocluster_consul in the previous commit that demonstrated the problem clearly.